Move application of default tags to tracer.cpp
